Move HttpRequest.kt/HttpResponse.kt object in client package
